+++ boilerplate/bot.py
@@ -1,5 +1,4 @@
 from pystark import Stark
 
 
-if __name__ == "__main__":
-    Stark().activate()
+Stark().activate()